+++ src/test/java/samples/parser/SDAParser.java
@@ -5,9 +5,9 @@ import java.io.IOException;
 import java.io.Reader;
 import java.util.Stack;
 
-import be.baur.sda.Node;
+import be.baur.sda.DataNode;
 import be.baur.sda.serialization.Parser;
-import be.baur.sda.serialization.SyntaxException;
+import be.baur.sda.serialization.SDAParseException;
 
 /**
  * Alternative SDA parser (actually the first one I wrote, and lacking support
@@ -20,9 +20,9 @@ import be.baur.sda.serialization.SyntaxException;
  * the parser returns a node 'greeting', containing a node 'message' with a
  * value of "hello".<br>
  */
-public final class SDAParser implements Parser {
+public final class SDAParser implements Parser<DataNode> {
 
-	public Node parse(Reader input) throws IOException, SyntaxException {
+	public DataNode parse(Reader input) throws IOException, SDAParseException {
 
 		Tokenizer lexer = new Tokenizer(input);
 		Stack<Token> stack = new Stack<Token>();
@@ -33,7 +33,7 @@ public final class SDAParser implements Parser {
 		 * new context node. When all input has been parsed, the context
 		 * node (representing the root element) is returned.
 		 */
-		Node context = null;
+		DataNode context = null;
 
 		do {
 			token = lexer.getToken();
@@ -45,7 +45,7 @@ public final class SDAParser implements Parser {
 				if ( stack.empty() ) {
 					// a root element, check whether it is the first (and only)
 					if (context != null) 
-						throw new SyntaxException("too many root elements", lexer.getPos());
+						throw new SDAParseException("too many root elements", lexer.getPos());
 					stack.push(token); continue;
 				}
 
@@ -53,7 +53,7 @@ public final class SDAParser implements Parser {
 					stack.push(token); continue; // identifier of a child element, put on the stack
 				}
 				// cannot have identifiers without a context (except the root)
-				throw new SyntaxException("unexpected identifier \"" + token.value + "\"", lexer.getPos());
+				throw new SDAParseException("unexpected identifier \"" + token.value + "\"", lexer.getPos());
 			}
 
 			// we got a string, must be value to an element
@@ -64,11 +64,11 @@ public final class SDAParser implements Parser {
 					if (stack.peek().type == Tokenizer.IDENTIFIER) {
 
 						// pop identifier, create a simple node
-						Node e; String name = stack.pop().value;
+						DataNode e; String name = stack.pop().value;
 						try {
-							e = new Node(name, token.value);
+							e = new DataNode(name, token.value);
 						} catch (IllegalArgumentException x) {
-							throw new SyntaxException(x.getMessage(), lexer.getPos());
+							throw new SDAParseException(x.getMessage(), lexer.getPos());
 						}
 						
 						if (context == null) context = e;
@@ -80,7 +80,7 @@ public final class SDAParser implements Parser {
 					}
 				}
 				// cannot have a value without preceding identifier
-				throw new SyntaxException("value \"" + token.value + "\" has no identifier", lexer.getPos());
+				throw new SDAParseException("value \"" + token.value + "\" has no identifier", lexer.getPos());
 			}
 
 			// we got the start of a block, must be a complex element then
@@ -91,11 +91,11 @@ public final class SDAParser implements Parser {
 					if (stack.peek().type == Tokenizer.IDENTIFIER) {
 
 						// pop identifier, create a complex node
-						Node c; String name = stack.pop().value;
+						DataNode c; String name = stack.pop().value;
 						try {
-							c = new Node(name); c.add(null);
+							c = new DataNode(name); c.add(null);
 						} catch (IllegalArgumentException x) {
-							throw new SyntaxException(x.getMessage(), lexer.getPos());
+							throw new SDAParseException(x.getMessage(), lexer.getPos());
 						}
 						
 						stack.push(token); // push block start on the stack
@@ -108,7 +108,7 @@ public final class SDAParser implements Parser {
 					}
 				}
 				// cannot start a block without preceding identifier
-				throw new SyntaxException("block has no identifier", lexer.getPos());
+				throw new SDAParseException("block has no identifier", lexer.getPos());
 			}
 
 			// we got the end of a block, parent of context becomes context
@@ -118,28 +118,28 @@ public final class SDAParser implements Parser {
 
 					if (stack.peek().type == Tokenizer.BLOCK_START) {
 
-						Node parent = context.getParent();
+						DataNode parent = context.getParent();
 						if (parent != null)	context = parent;
 						stack.pop(); continue;
 					}
 				}
 				// cannot start a block without preceding identifier
-				throw new SyntaxException("unexpected block end", lexer.getPos());
+				throw new SDAParseException("unexpected block end", lexer.getPos());
 			}
 
 			// no more input, check if we are done
 			if (token.type == Tokenizer.EOF) {
 
 				if ( !stack.empty() ) // still tokens left to reduce
-					throw new SyntaxException("unexpected end of input", lexer.getPos());
+					throw new SDAParseException("unexpected end of input", lexer.getPos());
 
 				if ( context == null ) // no nodes created
-					throw new SyntaxException("input has no data", lexer.getPos());
+					throw new SDAParseException("input has no data", lexer.getPos());
 
 				return context; // should be the top level node
 			}
 			// should never be reached
-			throw new SyntaxException("impossible error", lexer.getPos());
+			throw new SDAParseException("impossible error", lexer.getPos());
 		} 
 		while (true);
 	}