+++ src/test/java/test/TestSDAParserAlt.java
@@ -2,14 +2,15 @@ package test;
 
 import java.io.IOException;
 import java.io.StringReader;
+import java.util.function.Function;
 
 import be.baur.sda.Node;
-import be.baur.sda.serialization.SyntaxException;
+import be.baur.sda.serialization.SDAParseException;
 import samples.parser.SDAParser;
 
 public final class TestSDAParserAlt {
 
-	private static SDAParser parser = new samples.parser.SDAParser();
+	private static SDAParser parser = new SDAParser();
 	
 	private static String samplesda = "addressbook {\r\n" + 
 			"	contact {\r\n" + 
@@ -26,50 +27,52 @@ public final class TestSDAParserAlt {
 	
 	public static void main(String[] args) throws Exception {
 
-		Node helloworld = (new SDAParser()).parse(new StringReader("greeting{message\"hello world\"}"));
+		Node helloworld = parser.parse(new StringReader("greeting{message\"hello world\"}"));
 		System.out.println(helloworld);
 
-		Test t = new Test(s -> {
+		Function<String, String> strfun = str -> {
 			try {
-				return parser.parse(new StringReader(s)).toString();
+				return parser.parse(new StringReader(str)).toString();
 			} catch (Exception e) {
-				return e.getMessage();
+				return ((SDAParseException)e).getLocalizedMessage();
 			}
-		});
+		};
+				
+		Test s = new Test(strfun);
+		Test f = new Test(strfun, "error at position ");
 
 		// test valid SDA
-		t.test("S01", "empty\"\"", "empty \"\"");
-		t.test("S02", "  empty  \"\"  ", "empty \"\"");
-		t.test("S03", "empty{}", "empty { }");
-		t.test("S04", "  empty  {  }  ", "empty { }");
-		t.test("S05", "_c1 { s_1 \"hello  world\" } ", "_c1 { s_1 \"hello  world\" }");
-		t.test("S06", "example \"The \\\\ is called a \\\"backslash\\\" in English.\"", "example \"The \\\\ is called a \\\"backslash\\\" in English.\"");
+		s.ts1("S01", "empty\"\"", "empty \"\"");
+		s.ts1("S02", "  empty  \"\"  ", "empty \"\"");
+		s.ts1("S03", "empty{}", "empty { }");
+		s.ts1("S04", "  empty  {  }  ", "empty { }");
+		s.ts1("S05", "_c1 { s_1 \"hello  world\" } ", "_c1 { s_1 \"hello  world\" }");
+		s.ts1("S06", "example \"The \\\\ is called a \\\"backslash\\\" in English.\"", "example \"The \\\\ is called a \\\"backslash\\\" in English.\"");
 		
 		/* test valid SDA */
-		String s = "SDA syntax violation at position ";
-		t.test("F01", "", s + "0: input has no data");
-		t.test("F02", "  ", s + "2: input has no data");
-		t.test("F03", "bad", s + "3: unexpected end of input");
-		t.test("F04", "2bad", s + "1: identifier cannot start with '2'");
-		t.test("F05", "b@d", s + "2: identifier cannot contain '@'");
-		t.test("F06", "\"string\"", s + "8: value \"string\" has no identifier");
-		t.test("F07", "_ \"string\"", s + "10: invalid node name (_)");
-		t.test("F08", "\"pending", s + "8: trailing or pending quote");
-		t.test("F09", "trailing\"", s + "9: trailing or pending quote");
-		t.test("F10", "{", s + "1: block has no identifier");
-		t.test("F11", "_ {", s + "3: invalid node name (_)");
-		t.test("F12", " }", s + "2: unexpected block end");
-		t.test("F13", "noright {", s + "9: unexpected end of input");
-		t.test("F14", "noleft }", s + "8: unexpected block end");
-		t.test("F15", "a{} b{}", s + "6: too many root elements");
-		t.test("F16", "a{ b{} } }", s + "10: unexpected block end");
-		t.test("F17", "a \"b\" c \"d\"", s + "8: too many root elements");
+		f.ts1("F01", "", "0: input has no data");
+		f.ts1("F02", "  ", "2: input has no data");
+		f.ts1("F03", "bad", "3: unexpected end of input");
+		f.ts1("F04", "2bad", "1: identifier cannot start with '2'");
+		f.ts1("F05", "b@d", "2: identifier cannot contain '@'");
+		f.ts1("F06", "\"string\"", "8: value \"string\" has no identifier");
+		f.ts1("F07", "_ \"string\"", "10: invalid node name (_)");
+		f.ts1("F08", "\"pending", "8: trailing or pending quote");
+		f.ts1("F09", "trailing\"", "9: trailing or pending quote");
+		f.ts1("F10", "{", "1: block has no identifier");
+		f.ts1("F11", "_ {", "3: invalid node name (_)");
+		f.ts1("F12", " }", "2: unexpected block end");
+		f.ts1("F13", "noright {", "9: unexpected end of input");
+		f.ts1("F14", "noleft }", "8: unexpected block end");
+		f.ts1("F15", "a{} b{}", "6: too many root elements");
+		f.ts1("F16", "a{ b{} } }", "10: unexpected block end");
+		f.ts1("F17", "a \"b\" c \"d\"", "8: too many root elements");
 		
 		// test performance
 		UnitTestPerformance<String> p = new UnitTestPerformance<String>(str -> {
 			try {
 				parser.parse(new StringReader(str));
-			} catch (SyntaxException | IOException e) {
+			} catch (IOException e) {
 				e.printStackTrace();
 			}
 		});