+++ RELEASE_NOTES.md
@@ -1,15 +1,30 @@
 # Release Notes
 
-## [2.1.1] - 2023-10-20
+## [2.2.0] - 2024-02-20
 
-- Added Node.remove().
-- Closed issues: #3, #6 and #9.
+As of this release, Node is refactored into an interface (Node), an abstract 
+class (AbstractNode) and a concrete class (DataNode). Also, the exception 
+classes have been thrown around a bit (pun intended).
+
+- `Removed` java.text.ParseException.
+- `Renamed` SyntaxException to SDAParseException.
+- `Replaced` SDA.parser() with SDA.parse().
+- Added ParseException.
+- Added ProcessingException.
+- Added Node.get(Predicate).
+- Added Result and Results.
+- Closed issues: #7.
 
 ## Compatibility
 
 - Requires at least Java 8.
 
 ## Previous releases
+
+### [2.1.1] - 2023-10-20
+
+- Added Node.remove().
+- Closed issues: #3, #6 and #9.
 
 ### [2.1.0] - 2023-09-29